Fix crash when XCB fails to return window geometry
For some reasons, using a DDS texture in my pipeline caused
a crash in Renderdoc. libXCB fails to return the geometry,
leading to a null pointer deref.

I am not familiar at all with libXCB, so not sure if there is a
better way to handle this failure, but at least this prevents a crash.
Was able to display the texture in the UI with this fix.

An alternative would be to return a boolean when we fail to gather
window output dimensions. This would require a change across all
drivers, and I am not sure of the side effects this could cause.
(As for now, if the window is not found, we silently return).
For this reason, I decided to go with a simple band-aid patch.