Improve robustness of fbconfig selection for cloned GL context (glX)
The glXGetConfig(..., GLX_FBCONFIG_ID, &fbcfgID) call does not work
reliably on many implementations. In this case, all available fbconfigs
are queried and the first one is chosen, but this might be incompatible
to the visual of the window, resulting in a failure when the cloned
context is made current for the first time (issue: #2898).

This adds a fallback to query the fbconfig ID from the GL context if
querying it from the VisualInfo failed.