Remove specification of QToolButton popupMode when there is no menu
Specifically, the open button in the Python shell window and the blue
channel (only) in the texture viewer had values specified even though
they do not have an actual menu associated with the buttons. For the
texture viewer, this was explicitly set to the default value (delayed
popup); for the Python shell window the value is custom. The popup
mode is only used "for tool buttons that has a menu set or contains
a list of actions".

These were added in early commits for each feature, probably by
accident: 0837e66939b264fb9995624664ea2df1a5c40ad1 for the texture
viewer and 2a4596e06a78d6a3d335469afccda7bfeef86665 for the Python
shell.