Remove UNIX check for building include-bin
At this point, all environments that build RenderDoc using CMake need to
build the include-bin utility for the host. Thus, the UNIX check is
pointless. Note that building RenderDoc for Windows does not use CMake.

The UNIX variable is "set to True when the target system is UNIX or
UNIX-like". Apart from Windows (and maybe Nintendo Switch), it will be
true for all targets. More importantly, it will also be true when
cross-compiling from Windows (to e.g. Android). This is a problem,
because the assumption that a program named c++ exists does not always
hold on Windows hosts. (This problem is addressed in the next commit.)

For context, the `if(UNIX)` check is from a time when there was a
separate implementation targetted at Windows, which was removed in
commit cb9449230a17d92d6be9c0c8c37a2a1f01676391 (which replaced
`if(WINDOWS) ... else(UNIX)` with just `if(UNIX)`; I am not sure if
`else(UNIX)` instead of `elseif(UNIX)` even worked correctly or not).
There, a fragile objcopy-based implementation was used for Unix.
That objcopy-based implementation was replaced with xxd in commit
c1f66de4133dc8658da3cdbfbfeb70d5d44521f2, and then with include-bin in
commit 864fdbe179b6c1f20cae7c753fb018bec74b5738. Cross-compiling was
implemented in 279aa72064efac17bc9acb3cf035b63775f8b870. However,
the `if(UNIX)` check was not relevant in any of these (a Cygwin
install is as likely to have `objcopy` as `c++` (though its result
probably wouldn't combine nicely with a Visual Studio build), while my
Git Bash MSYS2 install has `xxd` but not `c++` or `objcopy`).

This commit mainly exists to avoid whitespace changes in the following
commit.