Clean up 3 small non-behavioral issues with Vulkan pixel history
The first one is a simply a typo.

The second one seems to be an unnecessary value that doesn't make sense
(as blend factors are ignored when blendEnable is false, and setting the
source color blend factor to the destination color seems like an
uncommon situation (the equivalent to blendEnable being false is
source = 1, dest = 0). This does not seem to be a driver bug, as
the commit that added it (50438987df78d66d40a5eb75446b9f97896ce432) does
not mention that nor does it seem to work around other driver bugs.

The third one is that VulkanDebugManager::PixelHistoryCopyPixel was
declared but never defined. VulkanPixelHistoryCallback::CopyImagePixel
serves the same purpose.