Functional tests cannot be built in CMake >= 3.12
In development/CMakeLists.txt:477, the Python interpreter and libs are discovered. It uses the 'old' package discovery mechanism and the new depending on CMake's version.

However the old mechanism sets different variables to the new, so the new also sets the old vars for compatibility, however it is missing two variables.

The first missing var causes the Python minor version to be hardcoded to zero (renderdoccmd/CMakeLists.txt:9) and the second causes the later conditional (CMakeLists.txt:495) to be malformed.