Fix D3D12 pixel history lacking primIDs for draws without a PS.
* No rendertarget was bound to the primitive ID generating draw
because rts is empty in draws without a rendertarget bound in the base
draw. Setting rts[0] was presumably stomping memory instead in this
case.
* Included is a fix for validation layer error thrown when reproducing
the problem, in D3D12ColorAndStencilCallback during a Pre/PostDispatch
CopyPixel uses the value of m_SavedState but was not set. In the case
of rpFlags, the value was uninitialized because D3D12RenderState does
not include it in its constructor. That caused an invalid resource
transition to be issued and was fortunately caught by the debug layer,
otherwise it was thus far invisible when running without debug on.