Add fence-aware multi-queue record chunk serialisation
Instead of serialising captured chunks purely by ascending event ID, which fails
to correctly replay the 'D3D12_Multi_Wait_Before_Signal' test, this walks the
commands for all captured queues and blocks/defers serialising queue commands
that are subsequent to a blocked 'fence wait'.

When queue serialisation reaches a blocked 'wait' it cycles queues to try to
serialise the next unblocked queue command chunks.

It tracks fence signal values within the capture so that we can resume serialising
queue command chunks after a suitably signalled, previously blocked fence 'wait'.

This could result in deadlocking the serialisation if a queue waits on a fence that
is signalled outside the scope of what is captured.  To deal with this, we first
assume that 'waited on' fences that are never seen to be signalled in our capture
records are simply considered signalled so we'll never block on them.  Then, if we
detect a deadlock on any other fence waits, we try to conservatively emulate a
suitable fence signal outside the capture's scope (happens in D3D12_Sharing test).
If all else fails, we resort to the previous event-ID-ordered serialisation.