D3D12 Pixel History fix copy from texture to buffer
Use D3D12_PLACED_SUBRESOURCE_FOOTPRINT Offset parameter to get to within 16KB of the destination buffer byte offset.
Set the copy texture footprint to be a 1D Texture with a width to be equal to the remaining byte offset / pixel_byte_size + 1.
Use a destination X of the remaining byte offset / pixel_byte_size