Added DXIL helpers IsDXCNop(), IsSSA(), IsLLVMDebugCall()
IsDXCNop() checks for DXC Nop generated instructions, which take the form of a Load from a global variable called "dx.nothing.*"

IsSSA() checks if a DXIL::Value is an SSA value i.e. from an instruction, not a constant etc.

Store function name from Operation::Call in a const reference instead of making a copy