Make the bytecode to ShaderEntry cast work for compute shaders
The ShaderEntry *CS() function returned bogus data (at least for me)
which crashed RenderDoc in FetchRootSig when trying to extract the
root signature (when compute->pRootSignature is NULL).
It worked fine for other shaders so I looked for relevant differences
between Serialise_CreateComputePipelineState and
Serialise_CreateGraphicsPipelineState and I think I found one...
But I'm not sure about this change because I do not fully
understand what's happening with those ShaderEntry-casts.