Fix crash in DX11 unmap if map was called before frame capture
The problem can be reproduced by calling map of a DX11 buffer and then keeping that open until a frame is captured in which unmap is called.
The call to
```cpp
bool found = FindDiffRange(MapWrittenData, record->GetShadowPtr(ctxMapID, 1), len, s, e);
```
would crash as `record->GetShadowPtr(ctxMapID, 1)` returns nullptr and `FindDiffRange` does not guard against this.