Support for viewing decoded ASTC in RenderDoc TextureViewer
Adds support for viewing decoded ASTC stored in dds files. We store ASTC in dds and this path makes it convenient to verify the integrity of the content. Currently only supports ASTC blocks with same width and height. For the remaining ASTC formats, the same old message is still printed

== Code ==
- dds_readwrite.cpp
  - Check if we loaded ASTC and decode it using astc_dec (https://github.com/richgel999/astc_dec/tree/master)
  - TextureDescription is filled as regular RGBA8/RGBA8Srgb with a new flag ResourceFormat_ASTC_Decode which is used by TextureViewer to display the original format