Update memory type for some external images on replay
For cases where an external image is a regular image on replay and the
original memory type for external memory cannot be used for regular
memory. This can only be done for dedicated allocation (which some
handle types require), since we know which image the memory will be
bound to when vkAllocateMemory is called and memory type is specified.

On some drivers, for externally imported AHB backed images the allowed
memory type bits (reported via vkGetAndroidHardwareBufferPropertiesANDROID)
has no overlap with the memory type bits for a similar non-external
image (reported via vkGetImageMemoryRequirements). In those cases,
RenderDoc is currently not able to replay a capture. It would be very
helpful if RenderDoc is capable of updating memory type on replay.